requirements.txt
absl-py
astor
backports.weakref
cffi
enum34
funcsigs
futures
gast
gin-config
google-pasta
grpcio
h5py
Keras-Applications
Keras-Preprocessing
Markdown
mock
numpy
protobuf
pycparser
setuptools
six
tensorboard
tensorflow-estimator
tensorflow
termcolor
Werkzeug
wheel
wrapt
keras
